Fixing warning of narrowing conversion

+++ Common/ML/src/OrtInterface.cxx
@@ -169,7 +169,7 @@ std::vector<O> OrtModel::v2v(std::vector<I>& input, bool clearInput)
 template <class I, class O> // class I is the input data type, e.g. float, class O is the output data type, e.g. O2::gpu::OrtDataType::Float16_t from O2/GPU/GPUTracking/ML/convert_float16.h
 std::vector<O> OrtModel::inference(std::vector<I>& input)
 {
-  std::vector<int64_t> inputShape{input.size() / mInputShapes[0][1], mInputShapes[0][1]};
+  std::vector<int64_t> inputShape{(int64_t)(input.size() / mInputShapes[0][1]), mInputShapes[0][1]};
   std::vector<Ort::Value> inputTensor;
   inputTensor.emplace_back(Ort::Value::CreateTensor<O>(pImplOrt->memoryInfo, (reinterpret_cast<O*>(input)).data(), input.size(), inputShape.data(), inputShape.size()));
   // input.clear();
@@ -184,7 +184,7 @@ std::vector<O> OrtModel::inference(std::vector<std::vector<I>>& input)
 {
   std::vector<Ort::Value> inputTensor;
   for (auto i : input) {
-    std::vector<int64_t> inputShape{i.size() / mInputShapes[0][1], mInputShapes[0][1]};
+    std::vector<int64_t> inputShape{(int64_t)(i.size() / mInputShapes[0][1]), mInputShapes[0][1]};
     inputTensor.emplace_back(Ort::Value::CreateTensor<O>(pImplOrt->memoryInfo, (reinterpret_cast<O*>(i)).data(), i.size(), inputShape.data(), inputShape.size()));
   }
   // input.clear();
@@ -206,7 +206,7 @@ std::string OrtModel::printShape(const std::vector<int64_t>& v)
 template <>
 std::vector<float> OrtModel::inference<float, float>(std::vector<float>& input)
 {
-  std::vector<int64_t> inputShape{input.size() / mInputShapes[0][1], mInputShapes[0][1]};
+  std::vector<int64_t> inputShape{(int64_t)(input.size() / mInputShapes[0][1]), mInputShapes[0][1]};
   std::vector<Ort::Value> inputTensor;
   inputTensor.emplace_back(Ort::Value::CreateTensor<float>(pImplOrt->memoryInfo, input.data(), input.size(), inputShape.data(), inputShape.size()));
   // input.clear();
@@ -219,7 +219,7 @@ std::vector<float> OrtModel::inference<float, float>(std::vector<float>& input)
 template <>
 std::vector<float> OrtModel::inference<OrtDataType::Float16_t, float>(std::vector<OrtDataType::Float16_t>& input)
 {
-  std::vector<int64_t> inputShape{input.size() / mInputShapes[0][1], mInputShapes[0][1]};
+  std::vector<int64_t> inputShape{(int64_t)(input.size() / mInputShapes[0][1]), mInputShapes[0][1]};
   std::vector<Ort::Value> inputTensor;
   inputTensor.emplace_back(Ort::Value::CreateTensor<Ort::Float16_t>(pImplOrt->memoryInfo, reinterpret_cast<Ort::Float16_t*>(input.data()), input.size(), inputShape.data(), inputShape.size()));
   // input.clear();
@@ -232,7 +232,7 @@ std::vector<float> OrtModel::inference<OrtDataType::Float16_t, float>(std::vecto
 template <>
 std::vector<OrtDataType::Float16_t> OrtModel::inference<OrtDataType::Float16_t, OrtDataType::Float16_t>(std::vector<OrtDataType::Float16_t>& input)
 {
-  std::vector<int64_t> inputShape{input.size() / mInputShapes[0][1], mInputShapes[0][1]};
+  std::vector<int64_t> inputShape{(int64_t)(input.size() / mInputShapes[0][1]), mInputShapes[0][1]};
   std::vector<Ort::Value> inputTensor;
   inputTensor.emplace_back(Ort::Value::CreateTensor<Ort::Float16_t>(pImplOrt->memoryInfo, reinterpret_cast<Ort::Float16_t*>(input.data()), input.size(), inputShape.data(), inputShape.size()));
   // input.clear();
@@ -245,7 +245,7 @@ std::vector<OrtDataType::Float16_t> OrtModel::inference<OrtDataType::Float16_t, 
 template <>
 std::vector<OrtDataType::Float16_t> OrtModel::inference<float, OrtDataType::Float16_t>(std::vector<float>& input)
 {
-  std::vector<int64_t> inputShape{input.size() / mInputShapes[0][1], mInputShapes[0][1]};
+  std::vector<int64_t> inputShape{(int64_t)(input.size() / mInputShapes[0][1]), mInputShapes[0][1]};
   std::vector<Ort::Value> inputTensor;
   inputTensor.emplace_back(Ort::Value::CreateTensor<Ort::Float16_t>(pImplOrt->memoryInfo, reinterpret_cast<Ort::Float16_t*>(input.data()), input.size(), inputShape.data(), inputShape.size()));
   // input.clear();
@@ -260,7 +260,7 @@ std::vector<OrtDataType::Float16_t> OrtModel::inference<OrtDataType::Float16_t, 
 {
   std::vector<Ort::Value> inputTensor;
   for (auto i : input) {
-    std::vector<int64_t> inputShape{i.size() / mInputShapes[0][1], mInputShapes[0][1]};
+    std::vector<int64_t> inputShape{(int64_t)(i.size() / mInputShapes[0][1]), mInputShapes[0][1]};
     inputTensor.emplace_back(Ort::Value::CreateTensor<Ort::Float16_t>(pImplOrt->memoryInfo, reinterpret_cast<Ort::Float16_t*>(i.data()), i.size(), inputShape.data(), inputShape.size()));
   }
   // input.clear();
